Fix -Wshorten-64-to-32: random.hpp.

+++ src/random.hpp
@@ -127,6 +127,6 @@ namespace randomness
 				break;
 			}
 		}
-		return std::distance(first, it);
+		return static_cast<unsigned int>(std::distance(first, it));
 	}
 }
